+++ lib/src/printer/object/utils.dart
@@ -1,0 +1,11 @@
+/// Default string to print before the fields.
+const defaultBeforeFields = '{';
+
+/// Default string to print between field name and value.
+const defaultFieldNameSeparator = '=';
+
+/// Default string to print between fields themselves.
+const defaultFieldSeparator = ', ';
+
+/// Default string to print after the fields.
+const defaultAfterFields = '}';